+++ qt6/network/gen_qsslconfiguration.cpp
@@ -342,19 +342,37 @@ void QSslConfiguration_setDefaultConfiguration(QSslConfiguration* configuration)
 }
 
 bool QSslConfiguration_dtlsCookieVerificationEnabled(const QSslConfiguration* self) {
+#if QT_CONFIG(dtls)
 	return self->dtlsCookieVerificationEnabled();
+#else
+	bool _ret_unavailable;
+	return _ret_unavailable;
+#endif
 }
 
 void QSslConfiguration_setDtlsCookieVerificationEnabled(QSslConfiguration* self, bool enable) {
+#if QT_CONFIG(dtls)
 	self->setDtlsCookieVerificationEnabled(enable);
+#else
+	return;
+#endif
 }
 
 QSslConfiguration* QSslConfiguration_defaultDtlsConfiguration() {
+#if QT_CONFIG(dtls)
 	return new QSslConfiguration(QSslConfiguration::defaultDtlsConfiguration());
+#else
+	QSslConfiguration* _ret_unavailable;
+	return _ret_unavailable;
+#endif
 }
 
 void QSslConfiguration_setDefaultDtlsConfiguration(QSslConfiguration* configuration) {
+#if QT_CONFIG(dtls)
 	QSslConfiguration::setDefaultDtlsConfiguration(*configuration);
+#else
+	return;
+#endif
 }
 
 bool QSslConfiguration_handshakeMustInterruptOnError(const QSslConfiguration* self) {
