+++ qt6/webengine/gen_qwebenginepage.cpp
@@ -32,7 +32,6 @@
 #include <QWebEngineNewWindowRequest>
 #include <QWebEnginePage>
 #include <QWebEngineProfile>
-#include <QWebEngineQuotaRequest>
 #include <QWebEngineRegisterProtocolHandlerRequest>
 #include <QWebEngineScriptCollection>
 #include <QWebEngineSettings>
@@ -55,7 +54,6 @@ void miqt_exec_callback_QWebEnginePage_windowCloseRequested(intptr_t);
 void miqt_exec_callback_QWebEnginePage_featurePermissionRequested(intptr_t, QUrl*, int);
 void miqt_exec_callback_QWebEnginePage_featurePermissionRequestCanceled(intptr_t, QUrl*, int);
 void miqt_exec_callback_QWebEnginePage_fullScreenRequested(intptr_t, QWebEngineFullScreenRequest*);
-void miqt_exec_callback_QWebEnginePage_quotaRequested(intptr_t, QWebEngineQuotaRequest*);
 void miqt_exec_callback_QWebEnginePage_registerProtocolHandlerRequested(intptr_t, QWebEngineRegisterProtocolHandlerRequest*);
 void miqt_exec_callback_QWebEnginePage_fileSystemAccessRequested(intptr_t, QWebEngineFileSystemAccessRequest*);
 void miqt_exec_callback_QWebEnginePage_selectClientCertificate(intptr_t, QWebEngineClientCertificateSelection*);
@@ -860,17 +858,6 @@ void QWebEnginePage_connect_fullScreenRequested(QWebEnginePage* self, intptr_t s
 	MiqtVirtualQWebEnginePage::connect(self, static_cast<void (QWebEnginePage::*)(QWebEngineFullScreenRequest)>(&QWebEnginePage::fullScreenRequested), self, [=](QWebEngineFullScreenRequest fullScreenRequest) {
 		QWebEngineFullScreenRequest* sigval1 = new QWebEngineFullScreenRequest(fullScreenRequest);
 		miqt_exec_callback_QWebEnginePage_fullScreenRequested(slot, sigval1);
-	});
-}
-
-void QWebEnginePage_quotaRequested(QWebEnginePage* self, QWebEngineQuotaRequest* quotaRequest) {
-	self->quotaRequested(*quotaRequest);
-}
-
-void QWebEnginePage_connect_quotaRequested(QWebEnginePage* self, intptr_t slot) {
-	MiqtVirtualQWebEnginePage::connect(self, static_cast<void (QWebEnginePage::*)(QWebEngineQuotaRequest)>(&QWebEnginePage::quotaRequested), self, [=](QWebEngineQuotaRequest quotaRequest) {
-		QWebEngineQuotaRequest* sigval1 = new QWebEngineQuotaRequest(quotaRequest);
-		miqt_exec_callback_QWebEnginePage_quotaRequested(slot, sigval1);
 	});
 }
