+++ qt6/webengine/gen_qwebenginepage.h
@@ -44,7 +44,6 @@ class QWebEngineNavigationRequest;
 class QWebEngineNewWindowRequest;
 class QWebEnginePage;
 class QWebEngineProfile;
-class QWebEngineQuotaRequest;
 class QWebEngineRegisterProtocolHandlerRequest;
 class QWebEngineScriptCollection;
 class QWebEngineSettings;
@@ -79,7 +78,6 @@ typedef struct QWebEngineNavigationRequest QWebEngineNavigationRequest;
 typedef struct QWebEngineNewWindowRequest QWebEngineNewWindowRequest;
 typedef struct QWebEnginePage QWebEnginePage;
 typedef struct QWebEngineProfile QWebEngineProfile;
-typedef struct QWebEngineQuotaRequest QWebEngineQuotaRequest;
 typedef struct QWebEngineRegisterProtocolHandlerRequest QWebEngineRegisterProtocolHandlerRequest;
 typedef struct QWebEngineScriptCollection QWebEngineScriptCollection;
 typedef struct QWebEngineSettings QWebEngineSettings;
@@ -164,8 +162,6 @@ void QWebEnginePage_featurePermissionRequestCanceled(QWebEnginePage* self, QUrl*
 void QWebEnginePage_connect_featurePermissionRequestCanceled(QWebEnginePage* self, intptr_t slot);
 void QWebEnginePage_fullScreenRequested(QWebEnginePage* self, QWebEngineFullScreenRequest* fullScreenRequest);
 void QWebEnginePage_connect_fullScreenRequested(QWebEnginePage* self, intptr_t slot);
-void QWebEnginePage_quotaRequested(QWebEnginePage* self, QWebEngineQuotaRequest* quotaRequest);
-void QWebEnginePage_connect_quotaRequested(QWebEnginePage* self, intptr_t slot);
 void QWebEnginePage_registerProtocolHandlerRequested(QWebEnginePage* self, QWebEngineRegisterProtocolHandlerRequest* request);
 void QWebEnginePage_connect_registerProtocolHandlerRequested(QWebEnginePage* self, intptr_t slot);
 void QWebEnginePage_fileSystemAccessRequested(QWebEnginePage* self, QWebEngineFileSystemAccessRequest* request);
